Create Day Schedule View
I began implementing DayScheduleView. I added and configured a
UIScrollView to display the schedule view in the available viewport and
support scrolling the schedule up or down.

I created TimeView. TimeView will draw the schedule grid within the
DayScheduleView and is the view that is actually scrolled by the scroll
view.

I created the Test playground to let me verify that the control is
working.