Design DayScheduleView Using PaintCode
I created the visual design for the DayScheduleView using PaintCode. I
exported the style kit and added it to the Xcode project.

I updated DayScheduleView to set the background color of the
UIScrollView to the background color from the style kit.

I updated TimeView to use the generated PaintCode drawing method to
paint the time view.