Create Delegate and Data Source Protocols
I created the DayScheduleViewDelegate protocol. DayScheduleView will
call delegate methods for actions that the user performs involving the
DayScheduleView.

I created the DayScheduleViewDataSource protocol. DayScheduleView will
call the data source to get the appointments to display in the view.

I added the delegate and dataSource properties to the DayScheduleView
class.