Move Current Time Line
I added code to TimeView to create a timer that fires every second. I
added code to determine when the minute changes and to invalidate the
area of the TimeView that the current time line is being displayed in
order to redraw that portion of the view to show the current time.