Draw Current Time Line
I implemented CurrentTimeLayerDelegate to draw the current time as a
horizontal line with the current time.

I implemented a timer in TimeView that will update the current time line
when the time changes.

I renamed DayScheduleViewSettings to DayScheduleViewMetrics. The use of
"metrics" just sounded better. I was never a fan of "settings."