Expose DayScheduleViewDelegate to Objective-C
I was forced to modify DayScheduleViewDelegate in order to decorate it
with @objc attributes and expose the delegate to Objective-C. This was
needed to implement RxSwift/RxCocoa support for DayScheduleView.

I changed DayScheduleViewAppointment to expose it to Objective-C.

I converted DayScheduleViewAppointmentEKEventAdapter from a struct to a
class. This was required because of adding the Objective-C support.